+++ src/api/encoding.cc
