+++ src/node_internals.h
@@ -56,6 +56,7 @@ class NativeModuleLoader;
 namespace per_process {
 extern Mutex env_var_mutex;
 extern double prog_start_time;
+extern bool v8_is_profiling;
 }  // namespace per_process
 
 // Forward declaration
